Converted modular robot rotations back to radians. SDF convertion still uses 90 degree angles. error is thrown is angles are not 90 degree

+++ core/revolve2/core/modular_robot/slot.py
@@ -4,9 +4,9 @@ from .serialized import Serialized
 
 class Slot:
     module: Module
-    rotation: int
+    rotation: float
 
-    def __init__(self, module: Module, rotation: int):
+    def __init__(self, module: Module, rotation: float):
         self.module = module
         self.rotation = rotation
